Fix `tensors used as indices must be long, byte or bool tensors` error

+++ recbole/sampler/sampler.py
@@ -187,7 +187,7 @@ class AbstractSampler(object):
                         if v in used
                     ]
                 )
-        return torch.tensor(value_ids)
+        return torch.tensor(value_ids, dtype=torch.long)
 
 
 class Sampler(AbstractSampler):
